+++ src/components/index.js
@@ -2,3 +2,4 @@ export  * as landingPageComp  from "./landingPageComponents/";
 export * as authPageComp from "./authPageComponents";
 export * as utilComp from "./utilComponents";
 export { RequiresAuth } from "./RequiresAuth";
+export { SafeFallback } from "./SafeFallback";
